+++ webapp/src/sui.tsx
@@ -1133,47 +1133,3 @@ export class Loader extends UIElement<LoaderProps, {}> {
         </div>
     }
 }
-
-export interface WebCamProps {
-
-}
-
-export class WebCam extends UIElement<WebCamProps, {}> {
-    private stream: MediaStream;
-    private v: HTMLVideoElement;
-
-    componentDidMount() {
-        navigator.getUserMedia({ video: true },
-            stream => {
-                this.stream = stream;
-                this.v.srcObject = this.stream;
-                this.v.play();
-            },
-            err => {
-                this.stop();
-            });
-    }
-
-    componentWillUnmount() {
-    }
-
-    private stop() {
-        if (this.stream) {
-            if (this.stream.stop)
-                this.stream.stop();
-            const tracks = this.stream.getTracks();
-            if (tracks)
-                tracks.forEach(track => track.stop());
-            this.stream = undefined;
-            this.v.srcObject = undefined;
-        }
-    }
-
-    private handleVideoRef = (ref: any) => {
-        this.v = ref;
-    }
-
-    render() {
-        return <video ref={this.handleVideoRef} />
-    }
-}
