+++ webapp/src/webcam.tsx
@@ -1,0 +1,64 @@
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+import * as sui from "./sui";
+import * as core from "./core";
+
+export interface WebCamProps {
+
+}
+
+export class WebCam extends sui.UIElement<WebCamProps, {}> {
+    private stream: MediaStream;
+    private v: HTMLVideoElement;
+
+    componentDidMount() {
+        navigator.mediaDevices.enumerateDevices()
+            .then(devices => {
+                core.dialogAsync({
+                    header: lf("Choose a camera"),
+                    hideCancel: devices.length > 0,
+                    buttons: devices.filter(device => device.kind == "videoinput")
+                        .map(device => {
+                            return {
+                                label: device.label,
+                                onclick: () => {
+                                    navigator.mediaDevices.getUserMedia({
+                                        video: { deviceId: { exact: device.deviceId } },
+                                        audio: false
+                                    }).then(stream => {
+                                        this.stream = stream;
+                                        this.v.srcObject = this.stream;
+                                        this.v.play();
+                                    }, err => {
+                                        this.stop();
+                                    })
+                                }
+                            } as ModalButton;
+                        })
+                })
+            })
+    }
+
+    componentWillUnmount() {
+    }
+
+    private stop() {
+        if (this.stream) {
+            if (this.stream.stop)
+                this.stream.stop();
+            const tracks = this.stream.getTracks();
+            if (tracks)
+                tracks.forEach(track => track.stop());
+            this.stream = undefined;
+            this.v.srcObject = undefined;
+        }
+    }
+
+    private handleVideoRef = (ref: any) => {
+        this.v = ref;
+    }
+
+    render() {
+        return <video ref={this.handleVideoRef} />
+    }
+}
